Move DynamicModal and Loading to system folder
