+++ components/layouts/Layout.tsx
@@ -1,6 +1,6 @@
 // Suspense
 import { Suspense } from "react";
-import Loading from "components/Loading";
+import Loading from "components/system/Loading";
 
 // Routing
 import Link from "next/link";
