+++ pages/settings/accessibility.tsx
@@ -3,7 +3,7 @@ import type { ReactElement } from "react";
 
 // Suspense
 import { Suspense } from "react";
-import Loading from "components/Loading";
+import Loading from "components/system/Loading";
 
 // SEO
 import Head from "next/head";
