+++ public/sw.js
@@ -1,2 +1,2 @@
-if(!self.define){let e,s={};const a=(a,c)=>(a=new URL(a+".js",c).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(c,n)=>{const i=e||("document"in self?document.currentScript.src:"")||location.href;if(s[i])return;let r={};const t=e=>a(e,i),d={module:{uri:i},exports:r,require:t};s[i]=Promise.all(c.map((e=>d[e]||t(e)))).then((e=>(n(...e),r)))}}define(["./workbox-6316bd60"],(function(e){"use strict";importScripts("fallback-tj6NW1OgtFM3uS7BASxWj.js"),self.addEventListener("message",(e=>{e.data&&"SKIP_WAITING"===e.data.type&&self.skipWaiting()})),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/chunks/182-97d2d8110ad2f347.js",revision:"97d2d8110ad2f347"},{url:"/_next/static/chunks/182-97d2d8110ad2f347.js.map",revision:"959e04fecdadc628216590e3fcd8bfa5"},{url:"/_next/static/chunks/218-f4943c68af6e3ed4.js",revision:"f4943c68af6e3ed4"},{url:"/_next/static/chunks/218-f4943c68af6e3ed4.js.map",revision:"daaa7779a345cb7f238b36b158809d08"},{url:"/_next/static/chunks/33-ba442087ae4fa8f5.js",revision:"ba442087ae4fa8f5"},{url:"/_next/static/chunks/33-ba442087ae4fa8f5.js.map",revision:"8cd9f26c7d5ccea02af65de26728b60d"},{url:"/_next/static/chunks/505-b90375e6a71271f6.js",revision:"b90375e6a71271f6"},{url:"/_next/static/chunks/505-b90375e6a71271f6.js.map",revision:"2f9356969de4186913076926e0b169de"},{url:"/_next/static/chunks/612-f273bace8a2f8032.js",revision:"f273bace8a2f8032"},{url:"/_next/static/chunks/612-f273bace8a2f8032.js.map",revision:"a89d35ea24012326f27314cf2b95a05f"},{url:"/_next/static/chunks/863-b4be9154836c3cc1.js",revision:"b4be9154836c3cc1"},{url:"/_next/static/chunks/863-b4be9154836c3cc1.js.map",revision:"5efd2995144cdda580fe8bc22d564dcf"},{url:"/_next/static/chunks/996-9c91d1eae613ddeb.js",revision:"9c91d1eae613ddeb"},{url:"/_next/static/chunks/996-9c91d1eae613ddeb.js.map",revision:"f8146419d29fa019023de28afea2bc50"},{url:"/_next/static/chunks/framework-da8ad0d3df154e7f.js",revision:"da8ad0d3df154e7f"},{url:"/_next/static/chunks/framework-da8ad0d3df154e7f.js.map",revision:"e43122222b79391baf565e0b2cea680d"},{url:"/_next/static/chunks/main-dda75576ff8dcbb9.js",revision:"dda75576ff8dcbb9"},{url:"/_next/static/chunks/main-dda75576ff8dcbb9.js.map",revision:"8ea4d1b06bf6684eb000fbfab9a361f1"},{url:"/_next/static/chunks/pages/404-edfa50908ac1a144.js",revision:"edfa50908ac1a144"},{url:"/_next/static/chunks/pages/404-edfa50908ac1a144.js.map",revision:"d48b22f6149df08fe6a54d8e7266b452"},{url:"/_next/static/chunks/pages/500-c376035374f39c20.js",revision:"c376035374f39c20"},{url:"/_next/static/chunks/pages/500-c376035374f39c20.js.map",revision:"e89228147f7a4a3585f5cffa9f38bd8c"},{url:"/_next/static/chunks/pages/_app-dcb44fd365bce0db.js",revision:"dcb44fd365bce0db"},{url:"/_next/static/chunks/pages/_app-dcb44fd365bce0db.js.map",revision:"0e05fa7e95330eb1adfd3351acdc4655"},{url:"/_next/static/chunks/pages/_error-e2f2bfcd5131d4b7.js",revision:"e2f2bfcd5131d4b7"},{url:"/_next/static/chunks/pages/_error-e2f2bfcd5131d4b7.js.map",revision:"304eea13ca7ffb487de091a049524fe8"},{url:"/_next/static/chunks/pages/_offline-d8d47ac9da83ec0d.js",revision:"d8d47ac9da83ec0d"},{url:"/_next/static/chunks/pages/_offline-d8d47ac9da83ec0d.js.map",revision:"34c1c5e1236be44df3766d5e9b660cd4"},{url:"/_next/static/chunks/pages/about/%5Bslug%5D-f6e0da0578d29d0c.js",revision:"f6e0da0578d29d0c"},{url:"/_next/static/chunks/pages/about/%5Bslug%5D-f6e0da0578d29d0c.js.map",revision:"0fb12c8af73f677c45ebe640653179f7"},{url:"/_next/static/chunks/pages/browse/%5Bslug%5D-095447187df254bf.js",revision:"095447187df254bf"},{url:"/_next/static/chunks/pages/browse/%5Bslug%5D-095447187df254bf.js.map",revision:"3dfbe10f2da7953b6bd055a27b53677c"},{url:"/_next/static/chunks/pages/composer-78deb807c39fc55a.js",revision:"78deb807c39fc55a"},{url:"/_next/static/chunks/pages/composer-78deb807c39fc55a.js.map",revision:"6a8db9f191445ff43b45e63a3677de24"},{url:"/_next/static/chunks/pages/docs/%5Bslug%5D-eacc6daa6eb2448d.js",revision:"eacc6daa6eb2448d"},{url:"/_next/static/chunks/pages/docs/%5Bslug%5D-eacc6daa6eb2448d.js.map",revision:"d8aa027f2777fd9e140f909aed24d41c"},{url:"/_next/static/chunks/pages/index-3e4fa1cc08f22743.js",revision:"3e4fa1cc08f22743"},{url:"/_next/static/chunks/pages/index-3e4fa1cc08f22743.js.map",revision:"52da03d4195498458daf0a2c4a57f41b"},{url:"/_next/static/chunks/pages/settings-b7618063ac47103c.js",revision:"b7618063ac47103c"},{url:"/_next/static/chunks/pages/settings-b7618063ac47103c.js.map",revision:"345a6efc997e0877a12c99e487051f3d"},{url:"/_next/static/chunks/pages/settings/accessibility-f4dce4d7c61c8fbf.js",revision:"f4dce4d7c61c8fbf"},{url:"/_next/static/chunks/pages/settings/accessibility-f4dce4d7c61c8fbf.js.map",revision:"1028d7424cf55e155aac1ff8fddaa70f"},{url:"/_next/static/chunks/pages/settings/general-d0224df60a81620e.js",revision:"d0224df60a81620e"},{url:"/_next/static/chunks/pages/settings/general-d0224df60a81620e.js.map",revision:"cff7f767efb9dc870b77798a51ec428f"},{url:"/_next/static/chunks/pages/settings/network-c78291132492448d.js",revision:"c78291132492448d"},{url:"/_next/static/chunks/pages/settings/network-c78291132492448d.js.map",revision:"2a85bccdfc1e0e1160bc8514857a553b"},{url:"/_next/static/chunks/pages/settings/sharing-9e909185f2101249.js",revision:"9e909185f2101249"},{url:"/_next/static/chunks/pages/settings/sharing-9e909185f2101249.js.map",revision:"5abf568f495ef0d01a12f0aed217c29d"},{url:"/_next/static/chunks/pages/settings/storage-98a7d73682979335.js",revision:"98a7d73682979335"},{url:"/_next/static/chunks/pages/settings/storage-98a7d73682979335.js.map",revision:"698f5aafba6154e62f32a1779091c49a"},{url:"/_next/static/chunks/polyfills-5cd94c89d3acac5f.js",revision:"99442aec5788bccac9b2f0ead2afdd6b"},{url:"/_next/static/chunks/webpack-d13296a0cd7aff3f.js",revision:"d13296a0cd7aff3f"},{url:"/_next/static/chunks/webpack-d13296a0cd7aff3f.js.map",revision:"d2be422a478c50efc8d48e7ef2366ce7"},{url:"/_next/static/css/e6e39b401b232a51.css",revision:"e6e39b401b232a51"},{url:"/_next/static/css/e6e39b401b232a51.css.map",revision:"9060667fa8fe2e142ec54d3edfba8588"},{url:"/_next/static/media/public-sans-all-400-normal.ae05de44.woff",revision:"ae05de44"},{url:"/_next/static/media/public-sans-all-600-normal.07428ec5.woff",revision:"07428ec5"},{url:"/_next/static/media/public-sans-latin-400-normal.66c4eba4.woff2",revision:"66c4eba4"},{url:"/_next/static/media/public-sans-latin-600-normal.7e06741e.woff2",revision:"7e06741e"},{url:"/_next/static/media/public-sans-latin-ext-400-normal.fd422b1b.woff2",revision:"fd422b1b"},{url:"/_next/static/media/public-sans-latin-ext-600-normal.47b27dc8.woff2",revision:"47b27dc8"},{url:"/_next/static/tj6NW1OgtFM3uS7BASxWj/_buildManifest.js",revision:"326485172d84568cf27cd1be67d842e5"},{url:"/_next/static/tj6NW1OgtFM3uS7BASxWj/_middlewareManifest.js",revision:"fb2823d66b3e778e04a3f681d0d2fb19"},{url:"/_next/static/tj6NW1OgtFM3uS7BASxWj/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_offline",revision:"tj6NW1OgtFM3uS7BASxWj"},{url:"/brand-assets/apple-icon-180.png",revision:"b1c325af8acb8c7676c05fade2d84d97"},{url:"/brand-assets/apple-splash-1125-2436.jpg",revision:"ff8297242458772c38815e8751ed534b"},{url:"/brand-assets/apple-splash-1136-640.jpg",revision:"df18cb3ac06008aefd770f978119af67"},{url:"/brand-assets/apple-splash-1170-2532.jpg",revision:"8a19d649ab7cd717743d761dabdbd523"},{url:"/brand-assets/apple-splash-1242-2208.jpg",revision:"83654f982fc7f0cc220a3ad56419cde2"},{url:"/brand-assets/apple-splash-1242-2688.jpg",revision:"801257d9ef8e940da5baa42f259cd259"},{url:"/brand-assets/apple-splash-1284-2778.jpg",revision:"f6a95ee55103f5d1c4a567e5a17e7e80"},{url:"/brand-assets/apple-splash-1334-750.jpg",revision:"a93a838d296c5194e56a709e26bf1ab7"},{url:"/brand-assets/apple-splash-1536-2048.jpg",revision:"f45d62acc433af6bc35bcce6648a4b82"},{url:"/brand-assets/apple-splash-1620-2160.jpg",revision:"024fe6103162542f995291215362f169"},{url:"/brand-assets/apple-splash-1668-2224.jpg",revision:"daac8a2f71c128c99dac85368feb680d"},{url:"/brand-assets/apple-splash-1668-2388.jpg",revision:"86609023e71521c3c9f120ecbdf66654"},{url:"/brand-assets/apple-splash-1792-828.jpg",revision:"f8d4f7cdc5bbe08c716041612789887b"},{url:"/brand-assets/apple-splash-2048-1536.jpg",revision:"5e123ac11a2bb4eb56cae42025432729"},{url:"/brand-assets/apple-splash-2048-2732.jpg",revision:"f63e8d9654c0d4955dc67ab7750721e2"},{url:"/brand-assets/apple-splash-2160-1620.jpg",revision:"5f4d4b760237ec18e56c53b3cbe51de1"},{url:"/brand-assets/apple-splash-2208-1242.jpg",revision:"205321dce8048af6aadf70869d9a6a37"},{url:"/brand-assets/apple-splash-2224-1668.jpg",revision:"3dad3cca716cb683522b9d9b1460385c"},{url:"/brand-assets/apple-splash-2388-1668.jpg",revision:"17e02f8f7c036ef3405c1875e523fa2d"},{url:"/brand-assets/apple-splash-2436-1125.jpg",revision:"586851ca942f9184a49ef18344d11fed"},{url:"/brand-assets/apple-splash-2532-1170.jpg",revision:"07fae1dc109697879da0bc2075691d0f"},{url:"/brand-assets/apple-splash-2688-1242.jpg",revision:"c2cbcedcca9b09defda6487e8a0cb67a"},{url:"/brand-assets/apple-splash-2732-2048.jpg",revision:"51514bc5e92fc32742468184c875c7ee"},{url:"/brand-assets/apple-splash-2778-1284.jpg",revision:"9c3bbf7a09432e6a124db30aab351a1c"},{url:"/brand-assets/apple-splash-640-1136.jpg",revision:"d8e73142f747c63058299d7448a7f5af"},{url:"/brand-assets/apple-splash-750-1334.jpg",revision:"ed3ce4270923df024a0b3ec51835bd9f"},{url:"/brand-assets/apple-splash-828-1792.jpg",revision:"949bee8db0061e27d5c18f7411199e33"},{url:"/brand-assets/favicon-196.ico",revision:"0ab6bc21395b1b573b6d2a27793dc1c7"},{url:"/brand-assets/manifest-icon-192.png",revision:"9a5cfcdbb311512a46f4d748ac634c03"},{url:"/brand-assets/manifest-icon-512.png",revision:"eb5a1c44fba6029b5f629de762a735c5"},{url:"/json/demo copy 2.json",revision:"c26e9e213e7b68504c0ce747e4a9039e"},{url:"/json/demo copy.json",revision:"fa996305655e4a537e59b5828656162c"},{url:"/json/demo.json",revision:"89ed6b18b093d75834703576236bf911"},{url:"/json/test.json",revision:"4bed32ba27e26a92c2e747e817ed5355"},{url:"/manifest.json",revision:"53149ae646cff0058e467800c1a5be88"},{url:"/markdown/about/privacy.mdx",revision:"bb373493fce83eafa88ee2ee7a46041e"},{url:"/markdown/about/terms.mdx",revision:"11d46f43bb20653a7ff27cdcf432cf9d"},{url:"/markdown/docs/accessibility.mdx",revision:"009322b82586a6d2deecd87b47cde940"},{url:"/markdown/docs/composer.mdx",revision:"a069aef044fc475ce0c850ef771aad38"},{url:"/markdown/docs/errors.mdx",revision:"7edbe69860198c757b53f292d8afe73a"},{url:"/markdown/docs/getting-started.mdx",revision:"2e65e78880e1a048883f22c454e4c430"},{url:"/markdown/docs/keyboard-shortcuts.mdx",revision:"6d7fa9b00bd262b675b619e084c89592"},{url:"/markdown/docs/settings.mdx",revision:"af213524db3a694aa2c383ac2b659551"},{url:"/markdown/docs/sharing.mdx",revision:"e2e4c09afd571f136c265f79ace9a410"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:a,state:c})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s},{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET")}));
+if(!self.define){let e,s={};const a=(a,c)=>(a=new URL(a+".js",c).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(c,n)=>{const i=e||("document"in self?document.currentScript.src:"")||location.href;if(s[i])return;let r={};const t=e=>a(e,i),d={module:{uri:i},exports:r,require:t};s[i]=Promise.all(c.map((e=>d[e]||t(e)))).then((e=>(n(...e),r)))}}define(["./workbox-6316bd60"],(function(e){"use strict";importScripts("fallback-pYSaMfdYMNLQjocRMRytq.js"),self.addEventListener("message",(e=>{e.data&&"SKIP_WAITING"===e.data.type&&self.skipWaiting()})),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/chunks/182-97d2d8110ad2f347.js",revision:"97d2d8110ad2f347"},{url:"/_next/static/chunks/182-97d2d8110ad2f347.js.map",revision:"959e04fecdadc628216590e3fcd8bfa5"},{url:"/_next/static/chunks/218-1ef4e2b119089cd8.js",revision:"1ef4e2b119089cd8"},{url:"/_next/static/chunks/218-1ef4e2b119089cd8.js.map",revision:"418204e2e4fd186af3cfe61432d66140"},{url:"/_next/static/chunks/33-ba442087ae4fa8f5.js",revision:"ba442087ae4fa8f5"},{url:"/_next/static/chunks/33-ba442087ae4fa8f5.js.map",revision:"8cd9f26c7d5ccea02af65de26728b60d"},{url:"/_next/static/chunks/505-b90375e6a71271f6.js",revision:"b90375e6a71271f6"},{url:"/_next/static/chunks/505-b90375e6a71271f6.js.map",revision:"2f9356969de4186913076926e0b169de"},{url:"/_next/static/chunks/612-f273bace8a2f8032.js",revision:"f273bace8a2f8032"},{url:"/_next/static/chunks/612-f273bace8a2f8032.js.map",revision:"a89d35ea24012326f27314cf2b95a05f"},{url:"/_next/static/chunks/863-b4be9154836c3cc1.js",revision:"b4be9154836c3cc1"},{url:"/_next/static/chunks/863-b4be9154836c3cc1.js.map",revision:"5efd2995144cdda580fe8bc22d564dcf"},{url:"/_next/static/chunks/996-9c91d1eae613ddeb.js",revision:"9c91d1eae613ddeb"},{url:"/_next/static/chunks/996-9c91d1eae613ddeb.js.map",revision:"f8146419d29fa019023de28afea2bc50"},{url:"/_next/static/chunks/framework-da8ad0d3df154e7f.js",revision:"da8ad0d3df154e7f"},{url:"/_next/static/chunks/framework-da8ad0d3df154e7f.js.map",revision:"e43122222b79391baf565e0b2cea680d"},{url:"/_next/static/chunks/main-dda75576ff8dcbb9.js",revision:"dda75576ff8dcbb9"},{url:"/_next/static/chunks/main-dda75576ff8dcbb9.js.map",revision:"8ea4d1b06bf6684eb000fbfab9a361f1"},{url:"/_next/static/chunks/pages/404-edfa50908ac1a144.js",revision:"edfa50908ac1a144"},{url:"/_next/static/chunks/pages/404-edfa50908ac1a144.js.map",revision:"d48b22f6149df08fe6a54d8e7266b452"},{url:"/_next/static/chunks/pages/500-c376035374f39c20.js",revision:"c376035374f39c20"},{url:"/_next/static/chunks/pages/500-c376035374f39c20.js.map",revision:"e89228147f7a4a3585f5cffa9f38bd8c"},{url:"/_next/static/chunks/pages/_app-5efa9afe0c8f86d4.js",revision:"5efa9afe0c8f86d4"},{url:"/_next/static/chunks/pages/_app-5efa9afe0c8f86d4.js.map",revision:"516f89010850c0fa216d821e0d85b100"},{url:"/_next/static/chunks/pages/_error-e2f2bfcd5131d4b7.js",revision:"e2f2bfcd5131d4b7"},{url:"/_next/static/chunks/pages/_error-e2f2bfcd5131d4b7.js.map",revision:"304eea13ca7ffb487de091a049524fe8"},{url:"/_next/static/chunks/pages/_offline-d8d47ac9da83ec0d.js",revision:"d8d47ac9da83ec0d"},{url:"/_next/static/chunks/pages/_offline-d8d47ac9da83ec0d.js.map",revision:"34c1c5e1236be44df3766d5e9b660cd4"},{url:"/_next/static/chunks/pages/about/%5Bslug%5D-2f93b6cae9104042.js",revision:"2f93b6cae9104042"},{url:"/_next/static/chunks/pages/about/%5Bslug%5D-2f93b6cae9104042.js.map",revision:"94640d3118b9fbfcf5b506ca72de0582"},{url:"/_next/static/chunks/pages/browse/%5Bslug%5D-b36a9b8ea1ca5b2c.js",revision:"b36a9b8ea1ca5b2c"},{url:"/_next/static/chunks/pages/browse/%5Bslug%5D-b36a9b8ea1ca5b2c.js.map",revision:"4275e09f2e689c02a80b77c2d51809f7"},{url:"/_next/static/chunks/pages/composer-f0840c8464ead456.js",revision:"f0840c8464ead456"},{url:"/_next/static/chunks/pages/composer-f0840c8464ead456.js.map",revision:"4e86c836ba17710f337d5c9dd76387e8"},{url:"/_next/static/chunks/pages/docs-9b2e529bd81ca761.js",revision:"9b2e529bd81ca761"},{url:"/_next/static/chunks/pages/docs-9b2e529bd81ca761.js.map",revision:"58b290ed75856f42e57a7a4d721c4291"},{url:"/_next/static/chunks/pages/docs/%5Bslug%5D-1f328f38bac7c378.js",revision:"1f328f38bac7c378"},{url:"/_next/static/chunks/pages/docs/%5Bslug%5D-1f328f38bac7c378.js.map",revision:"84379fcfc26fa2c5e812fb615e0518b8"},{url:"/_next/static/chunks/pages/docs/tour-a8a6b9bc2c5ce408.js",revision:"a8a6b9bc2c5ce408"},{url:"/_next/static/chunks/pages/docs/tour-a8a6b9bc2c5ce408.js.map",revision:"89eeb10aacceac25c41515528ce11305"},{url:"/_next/static/chunks/pages/index-b37474f83d881c8b.js",revision:"b37474f83d881c8b"},{url:"/_next/static/chunks/pages/index-b37474f83d881c8b.js.map",revision:"a6f1720e8ccde9b2cbd80d1bc6095141"},{url:"/_next/static/chunks/pages/settings-1908a69bf007db8a.js",revision:"1908a69bf007db8a"},{url:"/_next/static/chunks/pages/settings-1908a69bf007db8a.js.map",revision:"9241dcfe8b305549a0f24f397b6e12b4"},{url:"/_next/static/chunks/pages/settings/accessibility-05d432aa2fd5d1d9.js",revision:"05d432aa2fd5d1d9"},{url:"/_next/static/chunks/pages/settings/accessibility-05d432aa2fd5d1d9.js.map",revision:"5089baac6c0f715aa35e5cb5077b8279"},{url:"/_next/static/chunks/pages/settings/general-b3c9bfe5fba4e4c5.js",revision:"b3c9bfe5fba4e4c5"},{url:"/_next/static/chunks/pages/settings/general-b3c9bfe5fba4e4c5.js.map",revision:"34b866e7c8b831d27bdb13fe30a526d2"},{url:"/_next/static/chunks/pages/settings/network-c7b3e7c4ec20726e.js",revision:"c7b3e7c4ec20726e"},{url:"/_next/static/chunks/pages/settings/network-c7b3e7c4ec20726e.js.map",revision:"906546edbd4f2ac9d549b516b1e1358f"},{url:"/_next/static/chunks/pages/settings/sharing-54ade9fae3593fcc.js",revision:"54ade9fae3593fcc"},{url:"/_next/static/chunks/pages/settings/sharing-54ade9fae3593fcc.js.map",revision:"606017a9ea72e8ca1681ed95ba3fb31e"},{url:"/_next/static/chunks/pages/settings/storage-98a7d73682979335.js",revision:"98a7d73682979335"},{url:"/_next/static/chunks/pages/settings/storage-98a7d73682979335.js.map",revision:"698f5aafba6154e62f32a1779091c49a"},{url:"/_next/static/chunks/polyfills-5cd94c89d3acac5f.js",revision:"99442aec5788bccac9b2f0ead2afdd6b"},{url:"/_next/static/chunks/webpack-d13296a0cd7aff3f.js",revision:"d13296a0cd7aff3f"},{url:"/_next/static/chunks/webpack-d13296a0cd7aff3f.js.map",revision:"d2be422a478c50efc8d48e7ef2366ce7"},{url:"/_next/static/css/e6e39b401b232a51.css",revision:"e6e39b401b232a51"},{url:"/_next/static/css/e6e39b401b232a51.css.map",revision:"9060667fa8fe2e142ec54d3edfba8588"},{url:"/_next/static/media/public-sans-all-400-normal.ae05de44.woff",revision:"ae05de44"},{url:"/_next/static/media/public-sans-all-600-normal.07428ec5.woff",revision:"07428ec5"},{url:"/_next/static/media/public-sans-latin-400-normal.66c4eba4.woff2",revision:"66c4eba4"},{url:"/_next/static/media/public-sans-latin-600-normal.7e06741e.woff2",revision:"7e06741e"},{url:"/_next/static/media/public-sans-latin-ext-400-normal.fd422b1b.woff2",revision:"fd422b1b"},{url:"/_next/static/media/public-sans-latin-ext-600-normal.47b27dc8.woff2",revision:"47b27dc8"},{url:"/_next/static/pYSaMfdYMNLQjocRMRytq/_buildManifest.js",revision:"ee8e020ee264c84614c4d54b30e39708"},{url:"/_next/static/pYSaMfdYMNLQjocRMRytq/_middlewareManifest.js",revision:"fb2823d66b3e778e04a3f681d0d2fb19"},{url:"/_next/static/pYSaMfdYMNLQjocRMRytq/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_offline",revision:"pYSaMfdYMNLQjocRMRytq"},{url:"/brand-assets/apple-icon-180.png",revision:"b1c325af8acb8c7676c05fade2d84d97"},{url:"/brand-assets/apple-splash-1125-2436.jpg",revision:"ff8297242458772c38815e8751ed534b"},{url:"/brand-assets/apple-splash-1136-640.jpg",revision:"df18cb3ac06008aefd770f978119af67"},{url:"/brand-assets/apple-splash-1170-2532.jpg",revision:"8a19d649ab7cd717743d761dabdbd523"},{url:"/brand-assets/apple-splash-1242-2208.jpg",revision:"83654f982fc7f0cc220a3ad56419cde2"},{url:"/brand-assets/apple-splash-1242-2688.jpg",revision:"801257d9ef8e940da5baa42f259cd259"},{url:"/brand-assets/apple-splash-1284-2778.jpg",revision:"f6a95ee55103f5d1c4a567e5a17e7e80"},{url:"/brand-assets/apple-splash-1334-750.jpg",revision:"a93a838d296c5194e56a709e26bf1ab7"},{url:"/brand-assets/apple-splash-1536-2048.jpg",revision:"f45d62acc433af6bc35bcce6648a4b82"},{url:"/brand-assets/apple-splash-1620-2160.jpg",revision:"024fe6103162542f995291215362f169"},{url:"/brand-assets/apple-splash-1668-2224.jpg",revision:"daac8a2f71c128c99dac85368feb680d"},{url:"/brand-assets/apple-splash-1668-2388.jpg",revision:"86609023e71521c3c9f120ecbdf66654"},{url:"/brand-assets/apple-splash-1792-828.jpg",revision:"f8d4f7cdc5bbe08c716041612789887b"},{url:"/brand-assets/apple-splash-2048-1536.jpg",revision:"5e123ac11a2bb4eb56cae42025432729"},{url:"/brand-assets/apple-splash-2048-2732.jpg",revision:"f63e8d9654c0d4955dc67ab7750721e2"},{url:"/brand-assets/apple-splash-2160-1620.jpg",revision:"5f4d4b760237ec18e56c53b3cbe51de1"},{url:"/brand-assets/apple-splash-2208-1242.jpg",revision:"205321dce8048af6aadf70869d9a6a37"},{url:"/brand-assets/apple-splash-2224-1668.jpg",revision:"3dad3cca716cb683522b9d9b1460385c"},{url:"/brand-assets/apple-splash-2388-1668.jpg",revision:"17e02f8f7c036ef3405c1875e523fa2d"},{url:"/brand-assets/apple-splash-2436-1125.jpg",revision:"586851ca942f9184a49ef18344d11fed"},{url:"/brand-assets/apple-splash-2532-1170.jpg",revision:"07fae1dc109697879da0bc2075691d0f"},{url:"/brand-assets/apple-splash-2688-1242.jpg",revision:"c2cbcedcca9b09defda6487e8a0cb67a"},{url:"/brand-assets/apple-splash-2732-2048.jpg",revision:"51514bc5e92fc32742468184c875c7ee"},{url:"/brand-assets/apple-splash-2778-1284.jpg",revision:"9c3bbf7a09432e6a124db30aab351a1c"},{url:"/brand-assets/apple-splash-640-1136.jpg",revision:"d8e73142f747c63058299d7448a7f5af"},{url:"/brand-assets/apple-splash-750-1334.jpg",revision:"ed3ce4270923df024a0b3ec51835bd9f"},{url:"/brand-assets/apple-splash-828-1792.jpg",revision:"949bee8db0061e27d5c18f7411199e33"},{url:"/brand-assets/favicon-196.ico",revision:"0ab6bc21395b1b573b6d2a27793dc1c7"},{url:"/brand-assets/manifest-icon-192.png",revision:"9a5cfcdbb311512a46f4d748ac634c03"},{url:"/brand-assets/manifest-icon-512.png",revision:"eb5a1c44fba6029b5f629de762a735c5"},{url:"/json/demo copy 2.json",revision:"c26e9e213e7b68504c0ce747e4a9039e"},{url:"/json/demo copy.json",revision:"fa996305655e4a537e59b5828656162c"},{url:"/json/demo.json",revision:"89ed6b18b093d75834703576236bf911"},{url:"/json/test.json",revision:"4bed32ba27e26a92c2e747e817ed5355"},{url:"/manifest.json",revision:"53149ae646cff0058e467800c1a5be88"},{url:"/markdown/about/privacy.mdx",revision:"bb373493fce83eafa88ee2ee7a46041e"},{url:"/markdown/about/terms.mdx",revision:"11d46f43bb20653a7ff27cdcf432cf9d"},{url:"/markdown/docs/accessibility.mdx",revision:"009322b82586a6d2deecd87b47cde940"},{url:"/markdown/docs/composer.mdx",revision:"9c0424499d164cc430494a7a9b2445da"},{url:"/markdown/docs/errors.mdx",revision:"7edbe69860198c757b53f292d8afe73a"},{url:"/markdown/docs/keyboard-shortcuts.mdx",revision:"6d7fa9b00bd262b675b619e084c89592"},{url:"/markdown/docs/settings.mdx",revision:"65bd18d09097dfcbe3fcb8b829bea5f1"},{url:"/markdown/docs/sharing.mdx",revision:"e2e4c09afd571f136c265f79ace9a410"},{url:"/markdown/docs/storage.mdx",revision:"fe8b550114ed8ae26167cad1b106d36e"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:a,state:c})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s},{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET")}));
 //# sourceMappingURL=sw.js.map
